Read class is now in bamops
